fix: isolate action schemas and refactor log detail page
