+++ lib/actions/logs.ts
@@ -5,17 +5,8 @@ import { revalidatePath } from 'next/cache'
 import { join } from 'path'
 import { mkdir, writeFile } from 'fs/promises'
 import { redirect } from 'next/navigation'
-import { z } from 'zod'
 import { prisma } from '../db'
-
-export const logSchema = z.object({
-  projectId: z.string().cuid(),
-  date: z.string().min(1),
-  weather: z.string().optional(),
-  crewCount: z.coerce.number().int().optional(),
-  workDone: z.string().min(1),
-  notes: z.string().optional(),
-})
+import { logSchema } from '../validators/logs'
 
 export async function createDailyLog(_: unknown, formData: FormData) {
   const { userId } = auth()
