+++ lib/validators/contacts.ts
@@ -1,0 +1,9 @@
+import { z } from 'zod'
+
+export const contactSchema = z.object({
+  type: z.enum(['CLIENT','CONTRACTOR','VENDOR','INSPECTOR','BURRITO_TRUCK','OTHER']),
+  name: z.string().min(1),
+  phone: z.string().optional(),
+  email: z.string().email().optional(),
+})
+
